Move sin and cos functions from utils to maths/utils
