+++ packages/modeling/src/maths/mat4/fromXRotation.js
@@ -1,4 +1,4 @@
-const { sin, cos } = require('../../utils/trigonometry')
+const { sin, cos } = require('../utils/trigonometry')
 
 /**
  * Creates a matrix from the given angle around the X axis.
