+++ packages/modeling/src/maths/mat4/fromZRotation.js
@@ -1,4 +1,4 @@
-const { sin, cos } = require('../../utils/trigonometry')
+const { sin, cos } = require('../utils/trigonometry')
 
 /**
  * Creates a matrix from the given angle around the Z axis.
