+++ packages/modeling/src/maths/mat4/rotate.js
@@ -1,4 +1,4 @@
-const { sin, cos } = require('../../utils/trigonometry')
+const { sin, cos } = require('../utils/trigonometry')
 
 const copy = require('./copy')
