+++ packages/modeling/src/maths/utils/index.d.ts
@@ -3,5 +3,6 @@ export { default as area } from './area'
 export { default as interpolateBetween2DPointsForY } from './interpolateBetween2DPointsForY'
 export { default as intersect } from './intersect'
 export { default as solve2Linear } from './solve2Linear'
+export { sin, cos } from './trigonometry'
 
 export as namespace utils
