+++ packages/modeling/src/maths/utils/index.js
@@ -7,7 +7,9 @@
 module.exports = {
   aboutEqualNormals: require('./aboutEqualNormals'),
   area: require('./area'),
+  cos: require('./trigonometry').cos,
   interpolateBetween2DPointsForY: require('./interpolateBetween2DPointsForY'),
   intersect: require('./intersect'),
+  sin: require('./trigonometry').sin,
   solve2Linear: require('./solve2Linear')
 }
