+++ packages/modeling/src/maths/utils/trigonometry.d.ts
