+++ packages/modeling/src/maths/utils/trigonometry.test.js
