+++ packages/modeling/src/utils/index.d.ts
@@ -5,6 +5,5 @@ export { default as fnNumberSort } from './fnNumberSort'
 export { default as insertSorted } from './insertSorted'
 export { default as radiusToSegments } from './radiusToSegments'
 export { default as radToDeg } from './radToDeg'
-export * from './trigonometry'
 
 export as namespace utils
