+++ packages/modeling/src/utils/index.js
@@ -6,12 +6,10 @@
  */
 module.exports = {
   areAllShapesTheSameType: require('./areAllShapesTheSameType'),
-  cos: require('./trigonometry').cos,
   degToRad: require('./degToRad'),
   flatten: require('./flatten'),
   fnNumberSort: require('./fnNumberSort'),
   insertSorted: require('./insertSorted'),
   radiusToSegments: require('./radiusToSegments'),
-  radToDeg: require('./radToDeg'),
-  sin: require('./trigonometry').sin
+  radToDeg: require('./radToDeg')
 }
